+++ inc/sh/prompt.h
@@ -1,0 +1,24 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   prompt.h                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: dpoveda- <me@izenynn.com>                  +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2021/11/13 12:13:46 by dpoveda-          #+#    #+#             */
+/*   Updated: 2021/11/13 12:14:18 by dpoveda-         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef PROMPT_H
+# define PROMPT_H
+
+/*** PROTOYPES ***/
+
+/* prompt/prompt.c */
+char	*prompt_read_input(void);
+
+/* prompt/prompt_init.c */
+void	init_prompt(void);
+
+#endif